% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotate_it}
\alias{annotate_it}
\title{Annotation of text with NCBO ontologies}
\usage{
annotate_it(queryText, apikey = yourAPIkey, ontologies = NULL,
  expand_mappings = FALSE, longest_only = FALSE)
}
\arguments{
\item{queryText}{character - the string to annotate}

\item{apikey}{character - your unique API key (see NCBO account information on NCBO website)}

\item{ontologies}{character - the acronym of the ontology you wish to use (e.g. "NCIT")}

\item{expand_mappings}{boolean - TRUE means that the following manual mappings will be used in annotation: UMLS, REST, CUI, OBOXREF}

\item{longest_only}{boolean - TRUE means that only the longest match for a given phrase will be returned}
}
\value{
list parsed JSON call from the API
}
\description{
Annotates a queryText with NCBO ontologies
}
\examples{
annotate_it("gata4")

}
