% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ontologies}
\alias{get_ontologies}
\title{Get all current NCBO ontologies}
\usage{
get_ontologies(apikey = yourAPIkey)
}
\arguments{
\item{apikey}{character - your unique API key (see NCBO account information on NCBO website)}
}
\value{
list parsed JSON call from the API
}
\description{
Returns all current NCBO ontologies and their metadata
}
\examples{
o <- get_ontologies()

}
