% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_it}
\alias{search_it}
\title{Search NCBO ontology terms}
\usage{
search_it(queryText, apikey = yourAPIkey, require_exact_match = TRUE,
  ontologies = NULL, roots_only = FALSE)
}
\arguments{
\item{queryText}{character - the search term}

\item{apikey}{character - your unique API key (see NCBO account information on NCBO website)}

\item{require_exact_match}{boolean - default (TRUE)}

\item{ontologies}{character - the acronym of the ontology you wish to use (e.g. "NCIT")}

\item{root_only}{boolean - default (FALSE) true means performs root node only search}
}
\value{
list parsed JSON call from the API
}
\description{
Searches NCBO ontology terms with queryText
}
\examples{
search_it("gata4")

}
