% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_key}
\alias{set_key}
\title{Set global variable for your API key for other API functions}
\usage{
set_key(yourAPIkey)
}
\arguments{
\item{yourAPIkey}{character - your unique API key (see NCBO account information on NCBO website)}
}
\value{
global variable yourAPIkey set to your unique API key
}
\description{
Sets global variable yourAPIkey
}
\examples{
set_key("xxxxx-xxxx-xxxx-xxxx")

}
